PortCMIS: final minor changes before release
git-svn-id: https://svn.apache.org/repos/asf/chemistry/portcmis/trunk@1745123 13f79535-47bb-0310-9956-ffa450edef68

+++ PortCMIS/data/DataImpl.cs
@@ -1094,7 +1094,7 @@ namespace PortCMIS.Data
                     {
                         throw new ArgumentException("Property '" + Id + "' is a Decimal property!");
                     }
-                    return value;
+                    return (decimal)value;
                 case PropertyType.Uri:
                     if (!(value is string))
                     {
